pyproject.toml
[project]
name = "command-launcher"
version = "0.1.0"
description = "A command launcher utility"
authors = [{ name = "Y. Unal", email = "yunal22@ogr.iu.edu.tr" }]
requires-python = ">=3.8"
dependencies = [
    "pygobject>=3.48.2",
    "pyyaml>=6.0.3",
]

[project.scripts]
command-launcher = "command_launcher.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest",
]
